Tests: Refactor client of the lazy_typecheck module into a shared input.
This allows the ModuleInterface test to attempt to build a client against the
emitted `.swiftinterface` which will help us ensure the emitted interface is
correct.
